Set up initial hybrid branch
This removes the existing Polymer 2 component and its tests, and scaffolds this repo for Polymer 2 hybrid components. The expectation here is that we won't ever realistically be able to merge `hybrid`->`master` directly anyway, since there will be a huge number of conflicts no matter what. Instead, this (hybrid) branch will exist as effectively its own entity, until such time as the LMS supports Polymer 2+. When that time occurs, the components that exist in `hybrid` will need to be ported over to Polymer 2+ and `master`.

This also required removing the existing `d2l-organization-name` component (and demo/tests), as it was Polymer 2.