US97555 Add future/ended/inactive support
This isn't the fully-correct implementation, as that will involve adding the status badge between the header and content. However, this is a first pass, which moves the future/ended/inactive status of the organization into the `d2l-organization-info` component. This displays the same information that was previously displayed in the image overlay in the `d2l-enrollment-card`, meaning that can now be removed.